Convert executables to rubygems-centric loading
This fixes an issue in Rubygems 2.4.8 (likely introduced in rubygems
2.4.5, but that hasn't been confirmed) where, when an application loads
a plugin which has a rubygems runtime dep back on the main application,
files from both the app's primary location and its rubygem would both
be loaded, resulting in constant redefinition warnings and errors.

Specifically, the error is observed with chef-client, when loading
chef-provisioning-aws, which has a runtime dep on "chef", when these
projects are all installed to a ChefDK omnibus package. Chef is
initially loaded from /opt/chefdk/embedded/apps/chef, but when
'chef/provisioning/aws_driver' is required, files are then loaded from
'/opt/chefdk/embedded/lib/ruby/gems/2.1.0/gems/chef-12.4.1/lib'. This
eventually fails on a SuperclassMismatch error, due to code like:

```
class Foo < Struct.new(:item)
end
```

This change means that all appbundled applications need to exist as
*both* a normal source tree checkout (i.e., a git clone) and as a gem.
For this reason, the integration tests are broken by this commit and
marked as pending. Also, this makes the source tree checkout somewhat
redundant, since it won't be used at runtime.