omit locking bundler
Bundler doesn't emit equality pins for bundler itself if its a runtime
dep of one of the gems it is bundling.  For now just skip it.  I
suppose we could assume that the bundler we ran to bundle install is
the bundler that will run so we could `bundle --version` and use that
one.  The perf hit from rubygems loading up bundler is probably low
enough not to worry too hard about this right now though...