Warn on unresolved specs
These produce horrible perf degredation in rubygems internally:

https://github.com/rubygems/rubygems/issues/2761

Due to appbundler this "should" not happen, but due to:

https://github.com/chef/appbundler/commit/f01aea099ba3100e546a5ee315abc2a0e6d8878c

And due to ruby now shipping with bundler, we were getting two versions
of bundler and an unresolved spec, and therefore rubygems had
significant performance issues.  All that is required is correctly
activating bundler in the appbundler binstub and the spec becomes
resolved and the perf issues go away.

This commit just adds some warning about the issue so that people report
it, with a clear bit of text to not just ignore it.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>