force bundler activation
after we've activated all the other gems, force bundler activation.

since we're appbundling we've already activated all the gems and have
all the constraints loaded.  but bundler will not put bundler in a
Gemfile.lock due to bundler-ception issues.  we also do not want to
read and pin the version of bundler which is used to BUILD the lockfile
since that isn't necessarily the bundler version which omnibus will
install.

but once we're done we can force-activate bundler to resolve any
possible unresolved spec.

eventually we could be mildly smarter here if we need to and only do
this if ther are unresolved specs and if the unresolved spec is
bundler, but doing it all the time should be simple+safe.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>